Starting new MathGame Redux driven. Update MathGame to the test suite.
